Remove Isolator and :isolator argument
If a job wants to run out-of-process, it should manage that by itself.
We might provide a nice wrapper for a job in the future. For now only the case
where the job submits something into the execution context is not covered.

The problem is - process management is almost always bespoke, and it is difficult
to provide a process controller that will make things quit reliably, and on time.