Simplify eager loading for Rails
It is in fact the responsibility of the worker
process to force the eager loading of classes, so
we will take a two-pronged approach. First, we will
suggest the user to eager load in his worker process.
Second, we will actually perform the eager load
in the standard worker binary.