Ensure the provider thread gets killed forcibly first
If there is a configuration issue with SQS the `receive_messages`
call is going to hang. We need to account for this eventuality
and actually shoot the provider thread down so that the worker threads
can terminate normally. Potentially if these threads are submitting
and it also hangs that needs to be taken into account as well, but
that's for later.