Handle Seahorse errors with Retriable
The AWS SDK raises Seahorse::Client::NetworkingError when some-
thing goes wrong in the network. This was being caught and
handled by a custom retrying wrapper that retried up to 1000
times with 0.5 seconds in between. This commit switches to the
retriable gem, which uses a much nocer way of retries, with
exponential jittered backoff.