Better explain choice of number of retries for receives
Retriable has a maximum time during which retries are done. If
that time is reached it stops retrying even if max_retries has
not yet been reached. The default time (which is fine here) is
900 seconds = 15 minutes. For the receive I couldn't really be
bothered to calculate how much retries I would need to be
guarantueed to exceed 15 minutes with jitter and exponential
backoff so I just used a Really High Number (tm)