Stop overrevving receives
sqewer would become _very_ processor-aggressive if there were no jobs
in the queue to receive, because there were effectively no other threads
to schedule - so the producer thread (that pulls from SQS and sends to
the local cache) would be awoken multiple times per second, which leads
to a busywait. We have to be more gentle.