Permit local queues using SQLite (#10)
Permit local queues using SQLite

Vairous SQS emulators are in various stages of readiness.
For example, fake_sqs is processor hungry, and awsgo does not
support SendMessageBatch. Because we have an abstraction layer on
top of the actual message sends it's time to leverage it to permit
running _without_ any SQS emulators either

We also change the empty queue timeout when receive_messages returns
an empty array.

Sqewer would become _very_ processor-aggressive if there were no jobs
in the queue to receive, because there were effectively no other threads
to schedule - so the producer thread (that pulls from SQS and sends to
the local cache) would be awoken multiple times per second, which leads
to a busywait. We have to be more gentle.