Expose enqueue and enqueue_at both ways (#39)
ActiveJob in Rails 4 resolves the symbol value you give it
and then tries to call enqueue_* methods directly on what
got resolved. In Rails 5, first Rails will call .new on
what it resolved from the symbol and _then_ call enqueue
and enqueue_at on that what has gotten resolved. This means
that we have to expose these methods _both_ as class methods
and as instance methods.

This can be removed when we stop supporting Rails 4.

Also a good lesson that doing things in tests one way and suggesting
people use it in a different way in the documentation is probably
not such a good idea.