Raise immediately when message is too large
instead of waiting for the messagebox to be
flushed into the connection - as this would
create an error outside of the error tracking middleware,
preventing recovery and actually failing silently.

When a message turns out to be too large, fail immediately
and do not forward the mesasge to connection/messagebox.