Also retry on Aws::SQS::Errors::InternalError (#70)
Apparently our receiver can die if this happens, and previously we would not rescue it. This leaves the Sqewer worker in an invalid state, where the receiving thread is dead but the worker threads keep churning. While the worker process will terminate in this situation it is still much cheaper to make this retriable instead.

To terminate the worker thread, use `Thread#exit` instead of throwing as it is shorter and clearer to read.